k-animator-project/auth/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
azure-core==1.35.1
azure-identity==1.25.0
azure-keyvault-secrets==4.10.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.110.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
isodate==0.7.2
msal==1.33.0
msal-extensions==1.3.1
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.37.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
